style oracle tables a bit nicer
Fixes: https://github.com/iron-vault-plugin/iron-vault/issues/335
